create function to flag/report fraudulent car ad
