+++ SERVER/app.js
@@ -4,7 +4,7 @@ const logger = require("morgan");
 const bodyParser = require("body-parser");
 
 const app = express();
-const carRoutes =  require("./routes/car_route");
+const flagRoutes =  require("./routes/flag_route");
 
 
 app.use(bodyParser.json());
@@ -12,7 +12,7 @@ app.use(bodyParser.urlencoded({extended: false}));
 app.use(logger("dev"));
 
 
-app.use(("/car", carRoutes));
+app.use(("/", flagRoutes));
 
 module.exports = app;
