+++ SERVER/routes/flag_route.js
@@ -1,0 +1,12 @@
+const express = require("express");
+const router = express.Router();
+
+import Flag from '../controller/flag_controller';
+
+
+router.post('/api/v1/flag', Flag.create);
+
+
+
+module.exports = router;
+
